+++ uniform-config.js
@@ -1,0 +1,14 @@
+// Local uniform limit configuration. Change maxUniforms, then reload the unpacked extension.
+// You might as why this is in a separate file, this is to give a borderline room-temperature IQ
+// test to anyone trying to create infinity uniforms. So it's not just given OOB.
+// Since everyone asked so nicely, there will be a "soft" limit, emphasis on SOFT.
+window.TeamCraftersUniformConfig = {
+  maxUniforms: 10,
+};
+/**
+PS: I'm really bored. I've been out of work since June since the company went out of business.
+So I have a LOT of time to poke around now and find exploits in various things in the meantime!
+EA if you patch this, I'll likely just keep poking at it again and trying to see what I can do.
+If you want me to stop just find me a job to take up most of my time again lol -
+https://www.linkedin.com/in/jtrosclair/
+*/